+++ src/storage/src/cache/mod.rs
@@ -7,88 +7,7 @@ mod tracer;
 mod transcode;
 mod utils;
 
-use crate::liquid_array::LiquidArrayRef;
-use crate::sync::RwLock;
-use arrow::array::ArrayRef;
 pub use core::CacheStore;
 pub use policies::CachePolicy;
-use std::{fmt::Display, sync::Arc};
 pub use transcode::transcode_liquid_inner;
-pub use utils::{BatchID, CacheAdvice, CacheEntryID, ColumnAccessPath};
-
-/// States for liquid compressor.
-pub struct LiquidCompressorStates {
-    fsst_compressor: RwLock<Option<Arc<fsst::Compressor>>>,
-}
-
-impl std::fmt::Debug for LiquidCompressorStates {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "EtcCompressorStates")
-    }
-}
-
-impl Default for LiquidCompressorStates {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl LiquidCompressorStates {
-    /// Create a new instance of LiquidCompressorStates.
-    pub fn new() -> Self {
-        Self {
-            fsst_compressor: RwLock::new(None),
-        }
-    }
-
-    /// Get the fsst compressor.
-    pub fn fsst_compressor(&self) -> Option<Arc<fsst::Compressor>> {
-        self.fsst_compressor.read().unwrap().clone()
-    }
-}
-
-/// Cached batch.
-#[derive(Debug, Clone)]
-pub enum CachedBatch {
-    /// Cached batch in memory as Arrow array.
-    MemoryArrow(ArrayRef),
-    /// Cached batch in memory as liquid array.
-    MemoryLiquid(LiquidArrayRef),
-    /// Cached batch on disk as liquid array.
-    DiskLiquid,
-    /// Cached batch on disk as Arrow array.
-    DiskArrow,
-}
-
-impl CachedBatch {
-    /// Get the memory usage of the cached batch.
-    pub fn memory_usage_bytes(&self) -> usize {
-        match self {
-            Self::MemoryArrow(array) => array.get_array_memory_size(),
-            Self::MemoryLiquid(array) => array.get_array_memory_size(),
-            Self::DiskLiquid => 0,
-            Self::DiskArrow => 0,
-        }
-    }
-
-    /// Get the reference count of the cached batch.
-    pub fn reference_count(&self) -> usize {
-        match self {
-            Self::MemoryArrow(array) => Arc::strong_count(array),
-            Self::MemoryLiquid(array) => Arc::strong_count(array),
-            Self::DiskLiquid => 0,
-            Self::DiskArrow => 0,
-        }
-    }
-}
-
-impl Display for CachedBatch {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::MemoryArrow(_) => write!(f, "MemoryArrow"),
-            Self::MemoryLiquid(_) => write!(f, "MemoryLiquid"),
-            Self::DiskLiquid => write!(f, "DiskLiquid"),
-            Self::DiskArrow => write!(f, "DiskArrow"),
-        }
-    }
-}
+pub use utils::{BatchID, CacheAdvice, CacheEntryID, CachedBatch, ColumnAccessPath};
